+++ test/integ/util/system.py
@@ -563,6 +563,15 @@ class TestSystem(unittest.TestCase):
     self.assertEqual(home_dir, stem.util.system.expand_path('~%s' % username))
     self.assertEqual(os.path.join(home_dir, 'foo'), stem.util.system.expand_path('~%s/foo' % username))
 
+  def test_call_time_tracked(self):
+    """
+    Check that time taken in the call() function is tracked by SYSTEM_CALL_TIME.
+    """
+
+    initial = stem.util.system.SYSTEM_CALL_TIME
+    stem.util.system.call('sleep 0.5')
+    self.assertTrue(stem.util.system.SYSTEM_CALL_TIME - initial > 0.4)
+
   def test_set_process_name(self):
     """
     Exercises the get_process_name() and set_process_name() methods.